Make test_process_result_value rely on .return_value property
This commit modifies the way of getting the expected return value of
call to mock object representing Alias class. Passing the arguments used
in the call is not necessary to receive the return value, because the
same return value, stored in .return_value property, is returned
by all calls to an instance of Mock, even if the property has not been
set to a custom value.