Move imports from url_shortener.__init__ to run_server module
The imports being moved are modules event_handlers and views of package
url_shortener. They need to be imported to initialize views and event handlers
used by the application, and in the __ini__ module, they were imported
not on the top of the file, but after creation of the Flask app object.

Having imports not at the beginning of a module is contrary to PEP 8,
but, according to Flask documentation ([1]), "here it is actually fine".
Still, after dealing with some errors and reading code of other Flask
projects, I decided to move these imports into the module run_server, which
is responsible for running the application.

[1] http://flask.pocoo.org/docs/0.11/patterns/packages/