Fix ShortenURLTest.test_flashes_errors test method
After making shorten_url function flash all errors in the form, instead of
just the ones pertaining to its url field, the setup of form mock for the
test was no longer valid, making the test fail.

This commit fixes that by replacing an assignment to form_mock.url.errors
with an assignment to form_mock.errors.values.return_value.