Add init_random_factory class method to Alias
The method introduced in this commit is responsible for calcucalting and
assignment of _min_new_int and _max_new_int class attributes of Alias. These
attributes will be used by a method responsible for creating an instance
of Alias based on pseudo-randomly generated integer.

The init_random_factory method calculates the class attributes based on
parameters it receives. This makes it possible to configure random alias
generation outside of the module defining Alias class and other modules that
use (or might use) it. This makes it unnecessary for user-provided
configuration to have been loaded at the time of module initialization, which
is better for testing, as it shouldn't rely on user-provided application
config to be provided.