Make Alias use AliasType as its metaclass
AliasType.__new__ method creates a class object and calls its
init_random_factory class method, passing proper configuration values to it.
By making Alias class use AliasType as its metaclass, we make sure the class
method is called upon creation of the class. This solves problem of the Alias
class being in an invalid state until the method is called.

I had two other ideas for a solution for this problem:

1. Calling Alias.init_random_factory below class definition

I didn't like this solution, because it wasn't part of class definition, so
it could be missed if the code was moved.

2. Testing _min_new_int and _max_new_int where they are used: in
Alias.create_random, and raising an exception if they are unassigned.

I didn't like this solution, because it would mean the values would be tested
with each call to Alias.create_random. They aren't expected to be assigned
to None or other invalid values after their first assignment, so testing them
more than once (in case Alias.create_random was called more than once) would
be redundant.