Change handling of successful creation of a shortened URL
This commit introduces the following changes:
* alias is extracted from session with session.get with a default value None,
not by expecting its value to be assigned to its key and then handling
KeyError in case it is not.
* the existing shortened URL is fetched from database by calling
ShortenedURL.get method, not by calling get_or_404. The code simply assumes
the URL exists and accesses the expected attributes of object representing it
to display information on success of the request. If the URL does not exist,
None is returned and the attribute access on it raises AttributeError, which
is handled by the application as internal server error. So, in case the URL
expected to exist does not exist, it is no longer (wrongly) handled as an
instance of 404 error.
* the success message is created in python code and displayed using flash
function. Previous handling relied on the information template being a
special part of Twig template rendered by shorten_url function. The message
was displayed when a new shortened URL object was passed to template renderer.
It was a bad idea, motivated by presence of HTML tags in the message.