Add configuration loading to manage.py
Before commit 078b136, the code responsible for loading custom config was
part of url_shortener/__init__.py, so importing application object from
url_shortener package to manage.py worked, because the module had the access
to application object with valid database configuration. The commit moved
the config-loading code to url_shortener/__init__.py, so the old manage.py
could no longer work, which was noticed only today.

This commit fixes the problem with running manage.py script by adding to it
the code responsible for loading custom config. The solution is temporary:
after refactoring the application instantiation and configuration into
a factory function, it will be replaced with a call to that function.