Rename ShortenedURL to TargetURL and adjust variable names
This commit renames ShortenedURL class to TargetURL and also changes names
(or their parts) of variables and parameters so that they fit the new name
of the class.

The URL represented by instance of ShortenedURL class is only considered
shortened when its alias attribute has a value and it is persisted - so
that the application can provide a short alias URL. This is not the case for
newly created objects. Therefore, TargetURL is probably a more accurate name
than ShortenedURL.