Rename "alias" and "value" instance attributes
This commit renames "alias" and "value" attributes of instances of
TargetURL. They shouldn't be changed after creating a TargetURL and
successfully creating a short URL for it, because that change would
destroy a link of which some people know and which might be used by
them. Therefore, the attributes are being renamed to "_alias" and
"_value", respectively, to mark them conventionally as "private". They
are already used only by methods of TargetURL, so the change doesn't
break anything.