Add commit_changes_to_database function
This function provides a new way for shortening URLs by committing
changes to database. It separates adding new target URLs to current
session from shortening them by persisting them to database, which
means multiple URLs can now be committed in a single database
transaction.

It also provides a new way of handling integrity errors raised in
the process of committing changes. Detecting a number of these errors
that exceeds a configurable limit will no longer raise an exception
to be handled in a view by displaying an information to the user.
Instead, such situations will be handled without raising a new
exception: the operation of committing the whole transaction to
database will simply be retried until it's successful, and then
relevant information will be logged as a warning. This change is being
made for the following reasons:

* I feel like failure to create a URL due to the number of integrity
errors raised in the process would be too rare to justify writing
the code responsible for showing a user a special information dedicated
only to this kind of error, instead of just handling it like other
internal server errors or implementing alternative solutions

* instead of giving up and bothering a user with error information,
integrity errors can be handled automatically by retrying until
transaction is successful.

* if the error ever starts appearing significantly more frequently, it
is only a sign of another problem: there aren't enough unused alias
values allowed for current configuration. It can be handled only by
administrators by paying attention to logs and adjusting
the configuration when necessary.