Add TargetURLModule.get_target_url_class method
The responsibility of a dependency injection container is to provide
instances of objects and inject their dependencies into them.
In python, creating a class is, in fact, creating an instance of
a subclass of type by calling type with some parameters, so this
operation can be moved into a method of an Injector module.