Modify instantiation of target_url_class object
This commit makes TargetURLModule use its get_target_url_class method
to create an instance of target_url_class to be injected as
a dependency.

This commit also reintroduces eager creation of target_url_class.
The advantage of creating the class eagerly is valivation of its
user-specific configuration options the moment the application starts.