Refactor TargetURLModule.get_target_url_class
I used a direct call to type() to create TargetURL class knowing that
it was no different than creating a class using the syntax with 'class'
keyword. I did it this way just because I thought it shows better that
the function performs instantiation of an object, which is
the responsibility of dependency injection container. Still, I think
the alternative using 'class' keyword is more redable.