Add ShowURL view class
The purpose of the class is to replace render_preview, preview,
get_response and redirect_for functions in url_shortener.views module.

The functions will be replaced for the following reasons:

* although render_preview and get_response were introduced so that
preview and redirect_for wouldn't contain duplicate code, their usage
was still kind of repetitive because redirect_for and preview each
consisted of just a single call to get_response, just with different
arguments in each of the functions.

* I felt the functions were too small and too rarely used to deserve
being functions on their own.

* preview and redirect_for relied on injecting dependencies only to
inject them further. This is an antipattern because it's
the responsibility of dependency injection container.