Remove fragile and needless dependency injection
This commit removes automatic injection of dependencies into injector
module methods that aren't providers. This use injection is fragile,
because it makes correctness of configuring dependencies rely on
the order of calling "configure" methods of each instance of a module.

Automatic injection of an instance of Flask class into an injector
module methods is being replaced with non-automatic (not managed by
instance of Injector) injection into injector module constructors.

Creating a provider method for binding dependencies requiring
app object and automatically injecting an instance of Flask into it
could be a better alternative, but creating instances of such
dependencies wouldn't be eager, and I want them to be created eagerly
so that some configuration errors could be discovered before
the application starts handling requests.

Automatic injection of instance of IntegerAlias into another injector
module method is being replaced by instantiating the class inside
the method. No instance of IntegerAlias will be needed anywhere else.

The commit also removes a provider of AliasAlphabet instance. With
the other changes introduced here, its automatic injection is
no longer needed.