Change the way success messages are displayed
This commit changes the code responsible for displaying messages in
views.shorten_url function and in shorten_url.html template. From
now on, the feedback on a short URL being successfully provided
consists of multiple flashed messages, displayed in individual list
items.

The commit also changes CSS styles in a way similar to changes applied
in previous commit to form error messages: the border and background
styles are now applied to whole list instead of its individual items,
and margins and padding are also adjusted. The effect looks similar to
what it looked like before, though, because of the feedback message
used to being displayed as a single list item.