Disable pylint invalid-name error in test modules
The names of test methods should generally be based on name of a tested
method (or function) and describe some condition of a test. I think it
is reasonable to tolerate test method names that are considered too
long by pylint, because if their length was restricted to 30
characters, the names of tested methods and functions would have to be
much shorter.