Fix a homoglyph replacement error
This commit fixes an error occuring when calling
AliasAlphabet.from_string method on a string containing an unsupported
character (that is: one that is not a part of the alphabet) with a
multiletter homoglyph consisting of supported characters. In such cases,
the function raised AliasValueError beacuse of the character not being
supported while the expected behaviour should be for the character to be
replaced by its supported equivalent.